requirements.txt
pycryptodome==3.9.7
pyperclip==1.7.0
python-dotenv==0.12.0
